refactor: 파일구조 정리 및 주석 제거
