+++ src/app/timer/page.tsx
@@ -1,12 +1,8 @@
-import Countdown from '../../containers/timer/timer-container';
+import TimerContainer from '../../containers/timer/timer-container';
 
 export const metadata = {
   title: 'Timer',
 };
 export default function Timer() {
-  return (
-    <div>
-      <Countdown />
-    </div>
-  );
+  return <TimerContainer />;
 }
